Changed dataloader.py to dataset.py. Also changed float32 to float16 during processing for space reduction. Added Data_readme and updated Data_Pipeline_Readme

+++ data_utils.py
@@ -154,7 +154,7 @@ def preprocess_episode(
             f"{len(frames)} frames vs {len(actions)} actions; expected actions = frames - 1"
         )
 
-    latents = encode_latents(frames, vae, device=device, batch_size=batch_size)
+    latents = encode_latents(frames, vae, device=device, batch_size=batch_size).half()
 
     latents_path = episode_out / "latents.pt"
     actions_out_path = episode_out / "actions.one_hot.pt"
